+++ okx/Account.py
@@ -366,3 +366,18 @@ class AccountAPI(OkxClient):
         if type != '':
             params['type'] = type
         return self._request_with_params(POST, BILLS_APPLY, params)
+
+    # Get GLP daily performance history for a program (BROK-2261)
+    def get_glp_historical_performance(self, program, begin='', end='', limit=''):
+        params = {'program': program}
+        if begin != '':
+            params['begin'] = begin
+        if end != '':
+            params['end'] = end
+        if limit != '':
+            params['limit'] = limit
+        return self._request_with_params(GET, GLP_HISTORICAL_PERFORMANCE, params)
+
+    # Get current-day + MTD GLP snapshot for all enrolled programs (BROK-2261)
+    def get_glp_today_performance(self):
+        return self._request_without_params(GET, GLP_TODAY_PERFORMANCE)